--- components/ConfigureAmplify.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import { initAmplify } from '../utils/amplify-init';
-
-initAmplify();
-
-export default function ConfigureAmplifyClientSide() {
-  return null;
-}
