+++ src/app/page.tsx
@@ -1,13 +1,12 @@
 'use client';
 
-export default function HomePage() {
+import React, { Suspense } from 'react';
+import Login from '../components/molecules/Login/Login';
+
+export default function Page() {
   return (
-    <div className='nhsuk-grid-row' data-testid='page-content-wrapper'>
-      <div className='nhsuk-grid-column-two-thirds'>
-        <h1 className='nhsuk-heading-xl' data-testid='page-heading'>
-          Hello World!
-        </h1>
-      </div>
-    </div>
+    <Suspense>
+      <Login />
+    </Suspense>
   );
 }
