+++ src/app/signout/page.tsx
@@ -2,17 +2,15 @@
 
 import React, { Suspense, useEffect, useState } from 'react';
 import { signOut } from '@aws-amplify/auth';
-import { Redirect } from './Redirect';
+import { Redirect } from '../../components/molecules/Redirect/Redirect';
 
 export default function Page() {
   const [signedOut, setSignedOut] = useState(false);
 
   useEffect(() => {
-    console.log('sign out use effect');
     if (!signedOut) {
       signOut()
         .then(() => setSignedOut(true))
-        .then(() => console.log('sign out complete'))
         .catch((error) => console.error(error));
     }
   });
