+++ src/components/molecules/Header/Header.tsx
@@ -3,6 +3,7 @@ import concatClassNames from '@/src/utils/concat-class-names';
 import content from '@/src/content/content';
 import styles from './Header.module.scss';
 import { HeaderType } from './header.types';
+import LoginStatus from '../LoginStatus/LoginStatus';
 
 export function NHSNotifyHeader({ className, dataTestId }: HeaderType) {
   return (
@@ -50,12 +51,7 @@ export function NHSNotifyHeader({ className, dataTestId }: HeaderType) {
           )}
           id='content-header'
         >
-          {/* I am currently testing the link wrapper, this will change later when we implement auth as the link will change based on auth state */}
-          <div className='nhsuk-account__login' data-testid='login-link'>
-            <Link className='nhsuk-account__login--link' href='/'>
-              {content.components.headerComponent.links.logIn}
-            </Link>
-          </div>
+          <LoginStatus />
         </div>
       </div>
     </header>
