+++ src/components/molecules/Login/Login.tsx
@@ -14,12 +14,10 @@ function Login({
 }) {
   useEffect(() => {
     if (user && redirectPath) {
-      console.log('redirect', user, redirectPath);
       location.href = redirectPath;
     }
   }, [user, redirectPath]);
   if (redirectPath && user) {
-    console.log('redirecting', user, redirectPath);
     return (
       <h3>
         Redirecting to{' '}
@@ -29,17 +27,11 @@ function Login({
       </h3>
     );
   }
-  console.log('no redirect', user, redirectPath);
   return null;
 }
 
 const LoginWithAuthenticator = () => {
   const searchParams = useSearchParams();
-  console.log(
-    'searchParams',
-    searchParams.get('redirect'),
-    searchParams.entries().next()
-  );
   return withAuthenticator(Login, {
     variation: 'default',
     hideSignUp: true,
