+++ src/components/molecules/Redirect/Redirect.tsx
@@ -6,16 +6,9 @@ import { useSearchParams } from 'next/navigation';
 export const Redirect = () => {
   const searchParams = useSearchParams();
 
-  console.log(
-    'signout searchParams',
-    searchParams.get('redirect'),
-    searchParams.entries().next()
-  );
-
   const redirect = searchParams.get('redirect') ?? '/';
 
   useEffect(() => {
-    console.log('signout useeffect', redirect);
     location.href = redirect;
   }, [redirect]);
